allow to enable/disable curfew
